requirements.txt
albumentations~=1.3.1
matplotlib~=3.10.1
numpy~=2.2.4
Pillow~=11.1.0
PyYAML~=6.0.2
scikit-image~=0.25.2
scipy~=1.15.2
tensorboard~=2.19.0
tqdm~=4.67.1
torch~=2.6.0
torchio~=0.20.6
torchmetrics~=1.0.0
torchvision~=0.21.0
tabulate~=0.9.0
pandas~=2.2.3
shutil~=1.0.12